+++ demo/index.html
@@ -19,7 +19,6 @@
   <div id="content">
     <h1>If you can see this, something is broken (or JS is not enabled)!!.</h1>
   </div>
-  <script type="text-javascript" src="http://localhost:3000/webpack-dev-server.js"></script>
   <script type="text/javascript" src="assets/main.js"></script>
 </body>
 </html>
